+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/anthropic
 
+## 4.0.0-beta.26
+
+### Patch Changes
+
+- 90e2d8a: chore: fix unused vars not being flagged by our lint tooling
+- Updated dependencies [90e2d8a]
+  - @ai-sdk/provider-utils@5.0.0-beta.18
+
+## 4.0.0-beta.25
+
+### Patch Changes
+
+- 09bd27b: feat (provider/anthropic): add support for inference_geo provider option
+
 ## 4.0.0-beta.24
 
 ### Patch Changes