+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/devtools
 
+## 1.0.0-beta.12
+
+### Patch Changes
+
+- 90e2d8a: chore: fix unused vars not being flagged by our lint tooling
+
+## 1.0.0-beta.11
+
+### Patch Changes
+
+- e27ed76: feat(devtools): add new devtools integration for telemetry
+
 ## 1.0.0-beta.10
 
 ### Patch Changes