+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/gateway
 
+## 4.0.0-beta.47
+
+### Patch Changes
+
+- 90e2d8a: chore: fix unused vars not being flagged by our lint tooling
+- Updated dependencies [90e2d8a]
+  - @ai-sdk/provider-utils@5.0.0-beta.18
+
+## 4.0.0-beta.46
+
+### Patch Changes
+
+- 6b0a40d: feat (provider/gateway): add sort options
+
 ## 4.0.0-beta.45
 
 ### Patch Changes